pyproject.toml
[tool.poetry]
name = "curs-project"
version = "0.1.0"
description = ""
authors = ["Dmitry_Bezgubov <dmitrij-bezgubov@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.32.3"
pandas = "^2.2.3"
load-dotenv = "^0.1.0"
openpyxl = "^3.1.5"
tk = "^0.1.0"
pytest = "^8.3.3"
item = "^0.0.3"


[tool.poetry.group.dev.dependencies]
black = "^24.10.0"

flake8 = "^7.1.1"


[tool.poetry.group.lint.dependencies]
flake8 = "^7.1.1"
black = "^24.10.0"
isort = "^5.13.2"
mypy = "^1.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
# Максимальная длина строки
line-length = 119
# Файлы, которые не нужно форматировать
exclude = '''
(
  /(
      \.eggs         # Исключить несколько общих каталогов
    | \.git          # в корне проекта
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | dist
  )/
  | foo.py           # Также отдельно исключить файл с именем foo.py
                     # в корне проекта
)
'''